requirements.txt
pdfplumber
pytesseract
pdf2image
Pillow
opencv-python
rapidfuzz
